requirements.txt
pyjwt[crypto]>=2.9.0,<3.0.0
requests>=2.32.3,<3.0.0